requirements.txt
folium==0.18.0
numpy==2.1.3
Pillow==11.0.0
streamlit==1.37.1
streamlit_folium==0.23.2
streamlit_navigation_bar==3.3.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
